FE-53 Refactor : pages 내 스타일컴포넌트 분리
1. 스타일컴포넌트를 분리했습니다.
2. import 순서를 정리했습니다.
3. 불필요 한 console.log를 삭제했습니다.
