+++ src/pages/404/404.jsx
@@ -1,30 +1,9 @@
-import styled from 'styled-components';
-import Logo from '../../components/atoms/Logo/Logo';
-import Button from '../../components/atoms/Button/Button';
 import { useNavigate } from 'react-router-dom';
 
+import Logo from '../../components/atoms/Logo/Logo';
+import Button from '../../components/atoms/Button/Button';
 import Img404 from '../../assets/not-found-logo.png';
-
-const Wrapper = styled.div`
-  width: 100vw;
-  height: 100vh;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  button {
-    width: 120px;
-    margin-top: 20px;
-  }
-`;
-
-const Text404 = styled.p`
-  font-size: 14px;
-  font-weight: 400;
-  color: ${props => props.theme.color.gray.d4};
-  text-align: center;
-  margin-top: 30px;
-`;
+import { Text404, Wrapper } from './404style';
 
 const Page404 = () => {
   const navigate = useNavigate();
