+++ src/pages/EmailLogin/EmailLogin.jsx
@@ -1,8 +1,9 @@
 import axios from 'axios';
-import styled from 'styled-components';
+import { useState } from 'react';
+import { useNavigate } from 'react-router-dom';
+import { useRecoilValue, useSetRecoilState } from 'recoil';
+
 import { CommonWrapper } from '../../components/common/commonWrapper';
-import { Link, useNavigate } from 'react-router-dom';
-import { useRecoilState, useRecoilValue, useSetRecoilState } from 'recoil';
 import {
   accountnameValue,
   idValue,
@@ -13,35 +14,14 @@ import {
   userIntroValue,
   usernameValue,
 } from '../../atoms';
-import { useState } from 'react';
 import InputBox from '../../components/modules/InputBox/InputBox';
 import Button from '../../components/atoms/Button/Button';
-
-const FormWrapper = styled.div`
-  margin: 30px 34px;
-`;
-const SignUpLink = styled(Link)`
-  display: block;
-  font-weight: 400;
-  font-size: 12px;
-  line-height: 15px;
-  text-align: center;
-  margin-top: 25px;
-  color: ${props => props.theme.color.text.gray};
-`;
-const FormTitle = styled.h1`
-  text-align: center;
-  font-weight: 500;
-  font-size: 24px;
-  margin-top: 30px;
-`;
-const InputWrap = styled.form`
-  display: flex;
-  justify-content: space-between;
-  flex-direction: column;
-  gap: 16px;
-  margin: 30px 0;
-`;
+import {
+  FormTitle,
+  FormWrapper,
+  InputWrap,
+  SignUpLink,
+} from './emailLoginStyle';
 
 const EmailLogin = () => {
   const navigate = useNavigate();
@@ -52,7 +32,7 @@ const EmailLogin = () => {
   const [pwdErrMessage, setPwdMessage] = useState('');
   const [pwdNeedValid, setPwdNeedValid] = useState(true);
   const [isValid, setIsValid] = useState(false);
-  const [userData, setUserData] = useRecoilState(userDataAtom);
+  const setUserData = useSetRecoilState(userDataAtom);
   const setUsername = useSetRecoilState(usernameValue);
   const setAccountname = useSetRecoilState(accountnameValue);
   const setIntro = useSetRecoilState(userIntroValue);
