+++ src/pages/MyProfileAddProduct/MyProfileAddProduct.jsx
@@ -1,51 +1,24 @@
-import styled from 'styled-components';
+import { useRef } from 'react';
+import { useRecoilState } from 'recoil';
+import axios from 'axios';
+
 import Img from '../../components/atoms/Img/Img';
 import Label from '../../components/atoms/Label/Label';
 import Input from '../../components/atoms/Input/Input';
 import Button from '../../components/atoms/Button/Button';
 import { CommonWrapper } from '../../components/common/commonWrapper';
-import { useEffect, useRef, useState } from 'react';
-
 import UploadImg from '../../assets/img-button.png';
-import { useRecoilState, useRecoilValue } from 'recoil';
 import {
   productImgAtom,
   productLinkAtom,
   productNameAtom,
   productPriceAtom,
 } from '../../atoms';
-import axios from 'axios';
 import InputBox from '../../components/modules/InputBox/InputBox';
-
-const ProductForm = styled.form`
-  display: flex;
-  flex-direction: column;
-  padding: 30px 14.7%;
-  gap: 16px;
-  & label:not(:first-of-type) {
-    margin-top: 16px;
-  }
-  & label:first-of-type {
-    display: flex;
-    flex-direction: column;
-    position: relative;
-    img {
-      margin: 18px 0 14px;
-      cursor: pointer;
-    }
-    button {
-      position: absolute;
-      bottom: 30px;
-      right: 12px;
-    }
-  }
-`;
+import { ProductForm } from './addProductStyle';
 
 const MyProfileAddProduct = () => {
   const [productImgSrc, setProductImgSrc] = useRecoilState(productImgAtom);
-  const productName = useRecoilValue(productNameAtom);
-  const productPrice = useRecoilValue(productPriceAtom);
-  const productLink = useRecoilValue(productLinkAtom);
 
   const imgFileInput = useRef(null);
   // 버튼 클릭시 input 클릭 함수
