+++ src/pages/MyProfileEdit/profileEditStyle.js
@@ -1,0 +1,14 @@
+import styled from 'styled-components';
+import { CommonWrapper } from '../../components/common/commonWrapper';
+
+export const ProfileEditWrapper = styled(CommonWrapper)`
+  padding: 30px 34px;
+`;
+
+export const InputWrap = styled.form`
+  display: flex;
+  justify-content: space-between;
+  flex-direction: column;
+  gap: 16px;
+  margin: 30px 34px;
+`;
