+++ src/pages/PostUpload/PostUpload.jsx
@@ -1,53 +1,22 @@
 import axios from 'axios';
-import styled from 'styled-components';
-import Input from '../../components/atoms/Input/Input';
-import { CommonWrapper } from '../../components/common/commonWrapper';
-import Profile from '../../components/atoms/Profile/Profile';
-import UploadImg from '../../assets/upload-file.png';
 import { useCallback, useRef, useEffect } from 'react';
 import { useRecoilValue, useSetRecoilState, useRecoilState } from 'recoil';
+
+import Input from '../../components/atoms/Input/Input';
+import Profile from '../../components/atoms/Profile/Profile';
+import UploadImg from '../../assets/upload-file.png';
 import { postTxtValue, profileImgSrc, uploadImgSrcArray } from '../../atoms';
 import Img from '../../components/atoms/Img/Img';
 import Icon from '../../components/atoms/Icon/Icon';
-
-const UploadWrapper = styled(CommonWrapper)`
-  position: relative;
-`;
-
-const ProfileImgDiv = styled.div`
-  border-radius: ${props => props.theme.borderRadius.circle};
-`;
-
-const FlexDiv = styled.div`
-  display: flex;
-  padding: 20px 16px 16px;
-  gap: 13px;
-  min-height: 100px;
-`;
-
-const TextAreaDiv = styled.div`
-  width: 100%;
-`;
-
-const ImgUploadLabel = styled.label`
-  width: 50px;
-  height: 50px;
-  position: fixed;
-  border-radius: ${props => props.theme.borderRadius.circle};
-  bottom: 16px;
-  right: 16px;
-  cursor: pointer;
-`;
-
-const ItemBox = styled.ul`
-  display: flex;
-  gap: 10px;
-  margin-top: 10px;
-`;
-
-const SelectedImgLi = styled.li`
-  position: relative;
-`;
+import {
+  FlexDiv,
+  ImgUploadLabel,
+  ItemBox,
+  ProfileImgDiv,
+  SelectedImgLi,
+  TextAreaDiv,
+  UploadWrapper,
+} from './postUploadStyle';
 
 const PostUpload = () => {
   const setPostTxt = useSetRecoilState(postTxtValue);
