+++ src/pages/SplashScreen/SplashScreen.jsx
@@ -1,35 +1,8 @@
-import styled, { keyframes } from 'styled-components';
-import Logo from '../../components/atoms/Logo/Logo';
 import { useState } from 'react';
 
+import Logo from '../../components/atoms/Logo/Logo';
 import FullLogoImg from '../../assets/full-logo.png';
-
-const fadeInOut = keyframes`
-  0% {
-    opacity: 0;
-    transform: translateY(80px);
-  }
-  100% {
-    opacity: 1;
-    transform: translateY(0);
-  }
-`;
-
-const Wrapper = styled.div`
-  position: absolute;
-  top: 0;
-  left: 0;
-  width: 100vw;
-  height: 100vh;
-  background-color: #fff;
-  display: ${props => props.display};
-  justify-content: center;
-  align-items: center;
-  z-index: 20;
-  img {
-    animation: ${fadeInOut} 0.5s linear forwards;
-  }
-`;
+import { Wrapper } from './splashStyle';
 
 const SplashScreen = () => {
   const [display, setDisplay] = useState('flex');
